rewrite: update Spring Boot version, dependencies updates, major code refactoring, JDK upgrade
