--- backend/src/main/java/com/github/jaczerob/madamchuckle/annotations/ToontownAPILoader.java
@@ -1,17 +1,0 @@
-package com.github.jaczerob.madamchuckle.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-import org.springframework.core.annotation.AliasFor;
-import org.springframework.stereotype.Component;
-
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-@Component
-public @interface ToontownAPILoader {
-    @AliasFor(annotation = Component.class)
-    String value() default "";
-}
