--- backend/src/main/java/com/github/jaczerob/madamchuckle/configurations/ToontownClientConfig.java
@@ -1,31 +1,0 @@
-package com.github.jaczerob.madamchuckle.configurations;
-
-import java.util.Collections;
-
-import org.springframework.beans.factory.ObjectFactory;
-import org.springframework.boot.autoconfigure.http.HttpMessageConverters;
-import org.springframework.cloud.openfeign.support.SpringDecoder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.MediaType;
-import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
-
-import feign.Logger;
-import feign.codec.Decoder;
-
-@Configuration
-public class ToontownClientConfig {
-    @Bean
-    public Logger.Level feignLoggerLevel() {
-        return Logger.Level.FULL;
-    }
-
-    @Bean
-    @SuppressWarnings("deprecation")
-    public Decoder feignDecoder() {
-        MappingJackson2HttpMessageConverter converter = new MappingJackson2HttpMessageConverter();
-        converter.setSupportedMediaTypes(Collections.singletonList(MediaType.ALL));
-        ObjectFactory<HttpMessageConverters> objectFactory = () -> new HttpMessageConverters(converter);
-        return new SpringDecoder(objectFactory);
-    }
-}
