--- backend/src/main/java/com/github/jaczerob/madamchuckle/controllers/ToontownController.java
@@ -1,84 +1,0 @@
-package com.github.jaczerob.madamchuckle.controllers;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.github.jaczerob.madamchuckle.clients.ToontownClient;
-import com.github.jaczerob.madamchuckle.services.CacheService;
-import com.github.jaczerob.madamchuckle.toontown.loaders.fieldoffices.models.FieldOffices;
-import com.github.jaczerob.madamchuckle.toontown.loaders.news.models.News;
-import com.github.jaczerob.madamchuckle.toontown.loaders.population.models.Population;
-import com.github.jaczerob.madamchuckle.toontown.loaders.releasenotes.models.ReleaseNotes;
-import com.github.jaczerob.madamchuckle.toontown.loaders.status.models.Status;
-import com.github.jaczerob.madamchuckle.toontown.models.login.Authentication;
-import com.github.jaczerob.madamchuckle.toontown.models.login.LoginInfo;
-import com.github.jaczerob.madamchuckle.toontown.models.login.LoginResponse;
-import com.github.jaczerob.madamchuckle.toontown.models.login.QueueToken;
-
-@RestController
-@RequestMapping("/api/toontown")
-@CrossOrigin(origins = { "http://localhost:4200" }, allowCredentials = "true")
-public class ToontownController {
-    @Autowired private ToontownClient toontown;
-    @Autowired private CacheService cache;
-    
-    @GetMapping("/population")
-    public ResponseEntity<Population> getPopulation() {
-        return ResponseEntity.ok((Population) cache.get("POPULATION"));
-    }
-
-    @GetMapping("/news")
-    public ResponseEntity<News> getNews() {
-        return ResponseEntity.ok((News) cache.get("NEWS"));
-    }
-
-    @GetMapping("/releasenotes")
-    public ResponseEntity<ReleaseNotes> getReleaseNotes() {
-        return ResponseEntity.ok((ReleaseNotes) cache.get("RELEASE_NOTES"));
-    }
-
-    @GetMapping("/fieldoffices")
-    public ResponseEntity<FieldOffices> getFieldOffices() {
-        return ResponseEntity.ok((FieldOffices) cache.get("FIELD_OFFICES"));
-    }
-    
-    @GetMapping("/status")
-    public ResponseEntity<Status> getStatus() {
-        return ResponseEntity.ok((Status) cache.get("STATUS"));
-    }
-
-    @PostMapping(value="/login", consumes=MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<LoginResponse> login(@RequestBody LoginInfo loginInfo) {
-        LoginResponse loginResponse = toontown.login(loginInfo);
-        if (loginResponse.getSuccess().equals("false") || loginResponse.getSuccess().equals("partial"))
-            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(loginResponse);
-
-        return ResponseEntity.accepted().body(loginResponse);
-    }
-
-    @PostMapping(value="/updateQueue", consumes=MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<LoginResponse> updateQueue(@RequestBody QueueToken queueToken) throws InterruptedException {
-        LoginResponse loginResponse = toontown.updateQueue(queueToken);
-        if (loginResponse.getSuccess().equals("false"))
-            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(loginResponse);
-
-        return ResponseEntity.accepted().body(loginResponse);
-    }
-
-    @PostMapping(value="/authenticate", consumes=MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<LoginResponse> authenticate(@RequestBody Authentication authentication) throws InterruptedException {
-        LoginResponse loginResponse = toontown.authenticate(authentication);
-        if (loginResponse.getSuccess().equals("false"))
-            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(loginResponse);
-
-        return ResponseEntity.accepted().body(loginResponse);
-    }
-}
