--- backend/src/main/java/com/github/jaczerob/madamchuckle/repositories/DistrictPopulationRepository.java
@@ -1,7 +1,0 @@
-package com.github.jaczerob.madamchuckle.repositories;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.github.jaczerob.madamchuckle.entities.DistrictPopulationEntity;
-
-public interface DistrictPopulationRepository extends JpaRepository<DistrictPopulationEntity, Integer> {}
