--- backend/src/main/java/com/github/jaczerob/madamchuckle/repositories/TotalPopulationRepository.java
@@ -1,7 +1,0 @@
-package com.github.jaczerob.madamchuckle.repositories;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.github.jaczerob.madamchuckle.entities.TotalPopulationEntity;
-
-public interface TotalPopulationRepository extends JpaRepository<TotalPopulationEntity, Integer> {}
