--- backend/src/main/java/com/github/jaczerob/madamchuckle/toontown/loaders/Loader.java
@@ -1,7 +1,0 @@
-package com.github.jaczerob.madamchuckle.toontown.loaders;
-
-import com.github.jaczerob.madamchuckle.toontown.models.ToontownObject;
-
-public interface Loader<T extends ToontownObject> {
-    public T load();
-}
