--- backend/src/main/java/com/github/jaczerob/madamchuckle/toontown/loaders/fieldoffices/models/FieldOffice.java
@@ -1,20 +1,0 @@
-package com.github.jaczerob.madamchuckle.toontown.loaders.fieldoffices.models;
-
-import com.fasterxml.jackson.annotation.JsonSetter;
-
-import lombok.Data;
-
-@Data
-public class FieldOffice {
-
-    private String department;
-    private int difficulty;
-    private int annexes;
-    private boolean open;
-    private int expiring;
-
-    @JsonSetter("difficulty")
-    public void setDifficulty(int difficulty) {
-        this.difficulty = difficulty + 1;
-    }
-}
