--- backend/src/main/java/com/github/jaczerob/madamchuckle/toontown/loaders/news/models/News.java
@@ -1,17 +1,0 @@
-package com.github.jaczerob.madamchuckle.toontown.loaders.news.models;
-
-import com.github.jaczerob.madamchuckle.toontown.models.ToontownObject;
-
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-
-@Data
-@EqualsAndHashCode(callSuper=false)
-public class News extends ToontownObject {
-    private String title;
-    private int postId;
-    private String author;
-    private String body;
-    private String date;
-    private String image;
-}
