--- backend/src/main/java/com/github/jaczerob/madamchuckle/toontown/loaders/population/models/District.java
@@ -1,12 +1,0 @@
-package com.github.jaczerob.madamchuckle.toontown.loaders.population.models;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class District {
-    private String name;
-    private int population;
-    private String status;
-}
