--- backend/src/main/java/com/github/jaczerob/madamchuckle/toontown/loaders/releasenotes/models/ReleaseNotes.java
@@ -1,10 +1,0 @@
-package com.github.jaczerob.madamchuckle.toontown.loaders.releasenotes.models;
-
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-
-@Data
-@EqualsAndHashCode(callSuper=true)
-public class ReleaseNotes extends ReleaseNotesPartial {
-    private String body;
-}
