--- backend/src/main/java/com/github/jaczerob/madamchuckle/toontown/models/ToontownObject.java
@@ -1,3 +1,0 @@
-package com.github.jaczerob.madamchuckle.toontown.models;
-
-public class ToontownObject {}
