--- backend/src/main/java/com/github/jaczerob/madamchuckle/toontown/models/login/Authentication.java
@@ -1,9 +1,0 @@
-package com.github.jaczerob.madamchuckle.toontown.models.login;
-
-import lombok.Data;
-
-@Data
-public class Authentication {
-    private String appToken;
-    private String authToken;
-}
