--- backend/src/main/java/com/github/jaczerob/madamchuckle/toontown/models/login/LoginInfo.java
@@ -1,9 +1,0 @@
-package com.github.jaczerob.madamchuckle.toontown.models.login;
-
-import lombok.Data;
-
-@Data
-public class LoginInfo {
-    private String username;
-    private String password;
-}
