--- backend/src/main/java/com/github/jaczerob/madamchuckle/toontown/models/login/QueueToken.java
@@ -1,8 +1,0 @@
-package com.github.jaczerob.madamchuckle.toontown.models.login;
-
-import lombok.Data;
-
-@Data
-public class QueueToken {
-    private String queueToken;
-}
