+++ backend/src/main/java/dev/jaczerob/madamchuckle/toontown/models/ToontownObject.java
@@ -1,0 +1,4 @@
+package dev.jaczerob.madamchuckle.toontown.models;
+
+public class ToontownObject {
+}
