--- backend/src/test/java/com/github/jaczerob/madamchuckle/MadamchuckleApplicationTests.java
@@ -1,13 +1,0 @@
-package com.github.jaczerob.madamchuckle;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MadamchuckleApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
